@chore:添加一些配置

+++ carp-auth/carp-auth-security/src/main/java/com/github/rxyor/carp/auth/security/config/CarpAuthClientProperties.java
@@ -45,12 +45,12 @@ public class CarpAuthClientProperties {
     /**
      * access token 生命周期
      */
-    private Long accessTokenValiditySeconds = 3600L;
+    private Integer accessTokenValiditySeconds = 3600;
 
     /**
      * refresh token 生命周期
      */
-    private Long refreshTokenValiditySeconds = 30 * 86400L;
+    private Integer refreshTokenValiditySeconds = 30 * 86400;
 
     /**
      * 是否支持refresh token
